CLOUDP-64640: refactoring of k8sutils
