+++ scripts/dev/k8s_conditions.py
